[ruuvitag] Moved classes to internal package
All classes in bindings should be in internal package.
